Backend/requirements.txt
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
networkx==2.8.6
numpy==1.23.3
pandas==1.5.1
Pillow==9.2.0
pycparser==2.21
pycryptodome==3.15.0
pymongo==4.2.0
python-dateutil==2.8.2
pytz==2022.6
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
segtok==1.5.11
six==1.16.0
tabulate==0.8.10
threadpoolctl==3.1.0
urllib3==1.26.12
Werkzeug==2.2.2
yake==0.4.8
